Version using fcontext instead of pthreads.
NetHack writes more output to stdout than has place in the buffer,
so we cannot just make it do its thing and read from the pipe after
as it may block on writing and never get to the "yield" point.

Since we were planing to use the output to write a ttyrec, the plan
now is to override fflush, putchar and puts with #define tricks and
produce ttyrecs directly.