Introduce nle_ctx_t.
This way we send around pointers. We cannot call every function with
the right pointer, so we have a thread-local global pointer that is
set by nle_start and nle_step.

This is still broken of course as all of NetHack's globals are
merrily being read and written by NetHack.

Current plan to fix this:
  * Trust that all of NetHack's variable globals are in g.
  * Add g to nle_ctx_t.
  * Do #define magic to replace g with current_nle_ctx->g.